fixed typo and testing lines
The thing to think through here is really whether we'd like to encourage a student to load the full scanned book or include lightbox images for each page in the edition. Arguably, once it's done once it's all good (but that is only the case if referencing images stored at flickr or somewhere else...moving images between platforms could be a bit hassle.